[Fri Dec 27 2024]: feat: full chatters_app code added
In this commit complete development code is added. For front-end client folder and for back-end
server folder are added. Also docker-compose.yml file is added for container setup and to run the
appication using docker. This code is standalone on itself and can be run from terminal or by using
docker in local environment.

BREAKING CHANGE: Complete standalone development code added.